Notification request dialog implemented and improved daily notification logic.

+++ app/src/androidTest/java/com/randos/reminder/data/dao/TaskDaoTest.kt
@@ -58,7 +58,7 @@ class TaskDaoTest {
        val id = taskDao.insert(task)
 
         //When
-        taskDao.delete(task.copy(id = id))
+        taskDao.delete(task.copy(id = id.toInt()))
 
         //Then
         val tasks = taskDao.getTasks().first()
